Implement GetThreatSeverities()

+++ ads-api/Services/Analytic/Threat.cs
@@ -5,7 +5,7 @@ namespace Its.Ads.Api.Models
         public string ThreatName { get; set; }
         public string ThreatDetail { get; set; }
         public string? Serverity { get; set; }
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
         public double Percentage { get; set; }
         public string? MitrTechniqueName { get; set; }
         public string? MitrTacticName { get; set; }
